refactor out stdio
